server/requirements.txt
xml-marshaller==1.0.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
protobuf==3.20.*
fastavro
PyYAML
msgpack